Aggiunto i controllers mancanti

+++ app/controllers/users_controller.ts
@@ -1,6 +1,5 @@
 import type { HttpContext } from '@adonisjs/core/http'
 import User from '#models/user'
-
 export default class UsersController {
   public async index() {
     //GET - ritorna tutti gli utenti con il .all
@@ -8,19 +7,27 @@ export default class UsersController {
   }
 
   //POST - crea utenza
-  public async store({ request }: HttpContext) {
+  public async store({ request, response }: HttpContext) {
     const data = request.only(['full_name', 'email', 'password'])
-    return User.create(data)
+    const user = await User.create(data)
+    return response.created(user)
   }
 
   //GET - Otteniamo uno specifico user
-  public async show({ params }: HttpContext) {
-    return User.findByOrFail(params.id)
+  public async show({ params, response }: HttpContext) {
+    const user = User.findByOrFail(params.id)
+    if (!user) {
+      return response.notFound({ message: 'Utente non trovato' })
+    }
+    return user
   }
 
   //PUT - modifichiamo lo specifico user
-  public async edit({ params, request }: HttpContext) {
+  public async update({ params, request, response }: HttpContext) {
     const user = await User.findByOrFail(params.id)
+    if (!user) {
+      return response.notFound({ message: 'Utente non trovato' })
+    }
     const data = request.only(['full_name', 'email', 'password'])
     user.merge(data)
     await user.save()
@@ -28,8 +35,11 @@ export default class UsersController {
   }
 
   //DELETE - cancello uno specifico
-  public async destroy({ params }: HttpContext) {
+  public async destroy({ params, response }: HttpContext) {
     const user = await User.findByOrFail(params.id)
+    if (!user) {
+      return response.notFound({ message: 'Utente non trovato' })
+    }
     await user.delete()
     return { message: 'Utente cancellato con successo' }
   }
